check for unique usernames

+++ app/routes/join.tsx
@@ -13,10 +13,26 @@ import { Label } from '~/components/ui/label';
 
 import type { Route } from './+types/join';
 
+export type UsernameCheckResponse = {
+  exists: boolean;
+};
+
 export async function clientAction({ request }: Route.ActionArgs) {
   const formData = await request.formData();
   const username = formData.get('username');
   const room = formData.get('room');
+
+  const response = await fetch(
+    `/api/${room}/check-username?username=${username}`,
+    {
+      method: 'GET',
+    },
+  );
+  const json = (await response.json()) as UsernameCheckResponse;
+
+  if (json.exists) {
+    return { error: true };
+  }
 
   // store username
   sessionStorage.setItem(
@@ -31,7 +47,7 @@ export async function clientAction({ request }: Route.ActionArgs) {
   return redirect(`/room/${encodeURIComponent(String(room))}`);
 }
 
-export default function Join() {
+export default function Join({ actionData }: Route.ComponentProps) {
   return (
     <Form method="post">
       <Card className="w-[350px]">
@@ -53,7 +69,17 @@ export default function Join() {
             </div>
             <div className="flex flex-col space-y-1.5">
               <Label htmlFor="roomName">Room Name</Label>
-              <Input id="roomName" name="room" placeholder="Name of the room" />
+              <Input
+                id="roomName"
+                name="room"
+                placeholder="Name of the room"
+                className={actionData?.error ? 'border-destructive' : ''}
+              />
+              {actionData?.error && (
+                <p className="text-xs text-destructive">
+                  Username already taken
+                </p>
+              )}
             </div>
           </div>
         </CardContent>
